+++ src/features/trip/edit/components/Tabs/Expenses.tsx
@@ -10,8 +10,8 @@ import AppButton from "@features/ui/AppButton";
 import { useBreakpoints } from "@hooks/useBreakpoints";
 import useDialog from "@hooks/useDialog";
 
-import ExpenseDialog from "../../components/Expenses/ExpenseDialog";
-import type { Expense, ExpenseCategory, Trip } from "../../types";
+import ExpenseDialog from "../../../components/Expenses/ExpenseDialog";
+import type { Expense, ExpenseCategory, Trip } from "../../../types";
 import ContentCard from "./ContentCard";
 
 type ExpenseGroup = {
